# apps/web/lib/openapi/spec.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Kitty Learn HTTP API",
    "version": "1.0.0",
    "description": "Next.js Route Handlers under `/api`. Most routes require a Supabase session (browser cookies after login). Parent-scoped POST routes also require the authenticated user to own `childId`."
  },
  "servers": [{ "url": "/", "description": "Same host as the Next.js app" }],
...
  "paths": {
...
    "/api/leaderboard": {
      "get": {
        "tags": ["leaderboard"],
        "summary": "Global XP leaderboard (top learners)",
        "description": "Parent-only. Returns up to 50 children ranked by total XP (competition ranking). Includes `isMine` when the child belongs to the authenticated parent.",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": { "type": "integer", "minimum": 1, "maximum": 50, "default": 50 }
          }
        ],
        "responses": {
          "200": {
            "description": "Ranked leaderboard entries",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["entries"],
                  "properties": {
                    "entries": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/LeaderboardEntry" }
                    }
                  }
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "SessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "sb-access-token",
        "description": "Supabase Auth uses multiple cookies; in the browser, log in to the app first. Try it out must send credentials (cookies). Cookie names depend on the Supabase project ref."
      }
    },
    "responses": {
      "Unauthorized": {
        "description": "Not logged in",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorBody" },
            "example": { "error": "unauthorized" }
          }
        }
      },
      "Forbidden": {
        "description": "Authenticated but not parent of child",
        "content": { "text/plain": { "schema": { "type": "string" } } }
      },
...
      "InternalError": {
        "description": "Unexpected error",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorBody" },
            "example": { "error": "internal" }
          }
        }
      },
...
    },
    "schemas": {
      "ErrorBody": {
        "type": "object",
        "properties": {
          "error": { "type": "string" }
        }
      },
...
      "LeaderboardEntry": {
        "type": "object",
        "required": [
          "rank",
          "childId",
          "name",
          "age",
          "totalXp",
          "level",
          "streakDays",
          "isMine"
        ],
        "properties": {
          "rank": { "type": "integer", "minimum": 1 },
          "childId": { "type": "string", "format": "uuid" },
          "name": { "type": "string" },
          "age": { "type": "integer", "minimum": 3, "maximum": 10 },
          "avatarUrl": { "type": "string", "nullable": true },
          "totalXp": { "type": "integer", "minimum": 0 },
          "level": { "type": "integer", "minimum": 1 },
          "streakDays": { "type": "integer", "minimum": 0 },
          "isMine": { "type": "boolean" }
        }
      },
...
    }
  },
  "security": [{ "SessionCookie": [] }]
}